Learn how to use claims with jwt

+++ resources/store.py
@@ -1,6 +1,6 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from schemas import StoreSchema
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
 from db import db
@@ -52,6 +52,9 @@ class Store(MethodView):
     # delete a specific store with store_id
     @jwt_required
     def delete(self, store_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         store = StoreModel.query.get_or_404(store_id)
         db.session.delete(store)
         db.session.commit()
